Finished implementing a Transformer that can operate on matrices
Performs full self-attention as opposed to alternating row and column
attention.

Correctness of the code has not been checked, but it certainly is a close
approximation to the jax Transformer module.

Currently, the primary thing missing is turning off dropout at test time, but
that shouldn't be relevant to our project.